refactor packageing in design system
