+++ designsystem/src/main/java/com/sanaa/designsystem/design_system/theme/Theme.kt
@@ -2,10 +2,10 @@ package com.sanaa.designsystem.design_system.theme
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
-import com.sanaa.designsystem.design_system.color.LocalNovixColors
-import com.sanaa.designsystem.design_system.color.NovixColors
-import com.sanaa.designsystem.design_system.text_style.LocalNovixTextStyle
-import com.sanaa.designsystem.design_system.text_style.NovixTextStyle
+import com.sanaa.designsystem.design_system.theme.color.LocalNovixColors
+import com.sanaa.designsystem.design_system.theme.color.NovixColors
+import com.sanaa.designsystem.design_system.theme.text_style.LocalNovixTextStyle
+import com.sanaa.designsystem.design_system.theme.text_style.NovixTextStyle
 
 object Theme {
     val colors: NovixColors
